add weight_init.rs and fix relevant parts in some files
