+++ src/dlfs01/ch05/train_neural_net.rs
@@ -5,9 +5,9 @@
 #![allow(dead_code)]
 #![allow(unused_imports)]
 
-use crate::prelude::*;
 use super::two_layer_net::{Model, TwoLayerNet};
 use crate::dlfs01::common::choice::Choice;
+use crate::prelude::*;
 // use crate::dlfs01::dataset::{load_mnist, MNISTDataSetArray2};
 use crate::dlfs01::math::arange;
 use crate::dlfs01::MathFunc;
@@ -65,10 +65,16 @@ pub fn main() {
 
     // initialize a two-layer model
     println!("initialize a model...");
-    let weight_init: DistributionEnum = DistributionEnum::Uniform;
-    let weight_init_params: [FF; 2] = [-1.0, 1.0];
-    let mut network: TwoLayerNet<FF> =
-        TwoLayerNet::new(input_size, HIDDEN_SIZE, output_size, batch_axis, weight_init, &weight_init_params);
+    let weight_init: WeightInitEnum = WeightInitEnum::Normal;
+    let weight_init_params: FF = 1.0;
+    let mut network: TwoLayerNet<FF> = TwoLayerNet::new(
+        input_size,
+        HIDDEN_SIZE,
+        output_size,
+        batch_axis,
+        weight_init,
+        weight_init_params,
+    );
     network.print_detail();
     network.verbose = false;
