+++ src/prelude.rs
@@ -49,8 +49,16 @@ pub use crate::dlfs01::common::optimizers::optimizer_base::OptimizerBase;
 pub use crate::dlfs01::common::optimizers::optimizer::{AdaGrad, Adam, Nesterov, RMSprop, SGD};
 
 #[doc(no_inline)]
-pub use crate::dlfs01::common::param_initializers::{
+pub use crate::dlfs01::common::param_initializers::weight_init::*;
+
+#[doc(no_inline)]
+pub use crate::dlfs01::common::param_initializers::ndarray_init::{
     initialize_randomized_ndarray, DistributionEnum,
+};
+
+#[doc(no_inline)]
+pub use crate::dlfs01::common::param_initializers::weight_init::{
+    initialize_weight, WeightInitEnum,
 };
 
 #[doc(no_inline)]
